backend/requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
django-allauth==0.54.0
django-extensions==3.2.3
djangorestframework==3.14.0
dj-rest-auth==4.0.1


